requirements.txt
dj-database-url==0.4.0
Django==1.8.4
django-haystack==2.4.1
django-ipware==1.1.3
django-markdown-deux==1.0.5
django-markdownx==1.3
djorm-ext-pgfulltext==0.10
elasticsearch==2.2.0
gunicorn==19.4.5
Markdown==2.6.5
markdown2==2.3.0
Pillow==3.0.0
psycopg2==2.5.4
python-slugify==1.1.4
Unidecode==0.4.18
urllib3==1.14
wheel==0.24.0
whitenoise==2.0.6
Whoosh==2.7.2
